Adding repeating alarm

+++ app/src/main/java/com/italo/remind_me/broadcast/AlarmBroadcastReceiver.kt
@@ -7,11 +7,11 @@ import com.italo.remind_me.data.Alarm
 import com.italo.remind_me.data.AlarmRepository
 import com.italo.remind_me.notification.NotificationHelper
 import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import javax.inject.Inject
 
 @AndroidEntryPoint
 open class AlarmBroadcastReceiver : BroadcastReceiver() {
@@ -21,17 +21,19 @@ open class AlarmBroadcastReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
         if (intent.action != "sync") {
             println("chegou")
-            var alarm =
+            val alarm =
                 Alarm(
                     intent.getIntExtra("ALARM_ID", 0),
-                    intent.getStringExtra("ALARM_NAME"),
-                    21
+                    intent.getStringExtra("ALARM_NAME"), 0,
+                    intent.getBooleanExtra("ALARM_REPEAT", false)
                 )
             println("alarm no receiver $alarm")
             NotificationHelper.createNotification(context, alarm)
             CoroutineScope(Dispatchers.Main).launch {
                 withContext(Dispatchers.IO) {
-                    repository.deleteAlarm(alarm)
+                    if (alarm.repeat.not()) {
+                        repository.deleteAlarm(alarm)
+                    }
                 }
                 context.sendBroadcast(Intent("sync"))
             }
